Add first Api connection site and prediction

+++ prediction-engine/src/main/java/fr/lernejo/prediction/models/Temperature.java
@@ -1,3 +1,3 @@
 package fr.lernejo.prediction.models;
 
-public record Temperature(String date, int temperature){ }
+public record Temperature(String date, double temperature){ }
